+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 * argument shortcuts for LFTS forms: `nl` (don't print the language) and `ftr` (free translation)
 * columns for slides & other improvements
+* `[my smallcaps]{.smallcaps}` notation
 
 ### Fixed
 * topic index